fix numpy arrays ending up as float64 when reading in 8bit images

+++ src/anomalib/data/utils/image.py
@@ -355,7 +355,7 @@ def read_image(path: str | Path, as_tensor: bool = False) -> torch.Tensor | np.n
         return to_image(image) if as_tensor else np.array(image)
     else:
         image = Image.open(path).convert("RGB")
-        return to_dtype(to_image(image), torch.float32, scale=True) if as_tensor else np.array(image) / 255.0
+        return to_dtype(to_image(image), torch.float32, scale=True) if as_tensor else np.array(image, dtype=np.float32) / 255.0
 
 
 def read_mask(path: str | Path, as_tensor: bool = False) -> torch.Tensor | np.ndarray:
